README
This is a NetBeans project written in Java.

The Project contains code for a GUI program that solves the 0-1 Knapsack Problem using the two approaches below:
1.) Dynamic Programming approach
2.) Genetic Algorithm

This Project was prepared during my the second Semester at the Univerisity of the South Pacific while attending a course unit 
named CS214: Design and Analysis of Algorithms, for the year 2021. 

The directory and placement of files inside this repository follows the similar structure of how the NetBeans IDE stores it's project's
files in the NetBeans working directory (preferably a directory like C:\Users\myusername\Documents\MyNetbeansWorkspace\ , for example, in a Windows PC). 

I prefer NetBeans IDE because of it's ease and simplicity in designing and editing Java Swing GUI programs.
This program uses a lot of elements from the Java Swing Library, a GUI developement kit for designing GUI programs in Java.
There's also a new version of Java Swing, which is JavaFX.
If you want to design to design GUI programs in C++ or C#, you can use Visual Studio Community Edition for free.